Refactorizar para usar el mismo servicio get para traer listados y mostrarlos
